End-of-week progress on press release tests
- Changed environment from goldcorp test site to AES test site due to
AES site being live.
- Test now creates a new press release (with a date/time-stamped
headline and the date/time being 15 years from today) then publishes it
and
checks the live site to see if the headline of the latest press release
matches.
- Adding an image in the press release has been dropped for now due to
difficulties. Body text is added but is not checked.
- Due to delays in changes becoming live, the test will wait for up to
2 minutes while refreshing to see if the press release status changes
to “Live”.

TO COME: adding a document to a press release/editing a press release,
deleting a press release