Can now create, edit, and delete a press release
Test now creates a press release, checks it exists on live site, edits
the headline and checks that the live site updates accordingly, and
then deletes the press release and checks that it is no longer present
on the live site. While checking the live site, the test will wait for
up to two minutes (while refreshing every five seconds) for the desired
result. After two minutes of waiting the test will fail.
Tests have been verified to work on AES and Kinross sites in the
.newtest environment. I believe that this would also work on the
goldcorp site but cannot test that due to the site being set to “not
live” and “password-protected”. To switch between sites, you must
comment/uncomment lines in pageobjects/LoginPage/LoginPage.java and
util/EnvironmentType.java
Still missing from the test is the addition of images and attached
documents to the press release. These are missing due to difficulties
in getting webdriver to navigate the “Image Manager”/“File Explorer” to
select a file.