Image now included in press release
When creating the press release, the first .png file webdriver can find
is added to the body of the press release. Still need to have some way
of verifying that the image (and the body text) is present on the live
site. I am considering uploading a specific image to all the test
environments and then hardcoding it for easier verification.